+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - 2024-09-15 - fix incorrect artist associated with Gloria. Mistakenly set to Roy Orbison when it should've been set to Leon Rene.
 - 2024-09-15 - completely by accident managed to fix the setlist note numbering. Had to put a check in to make sure the note wasn't empty OR null, not just one or the either.
 - 2024-09-23 - added venue full text search column.
+- 2024-09-24 - added aliases and musicbrainz ids to 745/2101 venues.